feat(#302): Add 'as' and 'join' term classes with evaluation logic
